+++ src/main/java/net/bios/frenchvanilla/Commands.java
@@ -2,6 +2,7 @@ package net.bios.frenchvanilla;
 
 import net.bios.frenchvanilla.deathlock.DeathsCommand;
 import net.bios.frenchvanilla.home.HomeCommands;
+import net.bios.frenchvanilla.player_settings.PlayerSettingsCommand;
 import net.fabricmc.fabric.api.command.v1.CommandRegistrationCallback;
 
 public class Commands {
@@ -9,6 +10,7 @@ public class Commands {
         CommandRegistrationCallback.EVENT.register((dispatcher, dedicated) -> {
             DeathsCommand.register(dispatcher);
             HomeCommands.register(dispatcher);
+            PlayerSettingsCommand.register(dispatcher);
         });
     }
 }