+++ src/main/java/net/bios/frenchvanilla/Settings.java
@@ -16,6 +16,7 @@ public final class Settings extends Config implements ConfigContainer {
     public boolean restoreXp = true;
     public double deathKeyUnlockDistance = 10.0;
     public boolean homes = true;
+    public boolean playerSettings = true;
     public boolean timber = true;
     public int timberTasksPerTick = 54;
 }