+++ src/main/java/net/bios/frenchvanilla/timber/LogBreakEvent.java
@@ -11,6 +11,8 @@ import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
+import static net.bios.frenchvanilla.Components.PLAYER_SETTINGS;
+
 public class LogBreakEvent implements PlayerBlockBreakEvents.Before {
     public static void register() {
         PlayerBlockBreakEvents.BEFORE.register(new LogBreakEvent());
@@ -18,8 +20,9 @@ public class LogBreakEvent implements PlayerBlockBreakEvents.Before {
 
     @Override
     public boolean beforeBlockBreak(World world, PlayerEntity player, BlockPos pos, BlockState state, BlockEntity blockEntity) {
-        if (!FrenchVanilla.config.timber
-                || world.isClient
+        if (world.isClient
+                || !FrenchVanilla.config.timber
+                || !PLAYER_SETTINGS.get(player).settings().timber.value
                 || !TreeType.isLog(state.getBlock()))
             return true;